🌱 Refactor: Extract mock data and split Settings into sections (#181)
Phase 3A + 5A of the refactoring plan.

Phase 3A — Extract mock data:
- Moved inline mock security data from Security.tsx to src/mocks/securityData.ts
- Security.tsx reduced from 1309 to 1177 lines

Phase 5A — Split Settings.tsx into sections:
- Settings.tsx reduced from 960 to 86 lines (orchestrator only)
- Extracted 9 focused section components:
  - GitHubTokenSection (245 lines)
  - ThemeSection (215 lines)
  - AgentSection (129 lines)
  - TokenUsageSection (116 lines)
  - AccessibilitySection (107 lines)
  - ProfileSection (80 lines)
  - AISettingsSection (76 lines)
  - APIKeysSection (41 lines)
  - PermissionsSection (19 lines)
