+++ web/src/components/settings/sections/index.ts
@@ -1,0 +1,9 @@
+export { AISettingsSection } from './AISettingsSection'
+export { ProfileSection } from './ProfileSection'
+export { AgentSection } from './AgentSection'
+export { GitHubTokenSection } from './GitHubTokenSection'
+export { APIKeysSection } from './APIKeysSection'
+export { TokenUsageSection } from './TokenUsageSection'
+export { ThemeSection } from './ThemeSection'
+export { AccessibilitySection } from './AccessibilitySection'
+export { PermissionsSection } from './PermissionsSection'
